requirements.txt
alembic==1.7.5
black==21.11b1
blinker==1.4
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
PyMySQL==1.0.2
python-dotenv==0.19.2
regex==2021.11.10
SQLAlchemy==1.4.27
tomli==1.2.2
typing_extensions==4.0.0
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
